feat(runtime): tracing GC cycle collection & native RFC 6455 WebSocket client (#11)
* feat(runtime): implement tracing garbage collection and cycle detection

- Add mark-and-sweep cycle collector with O(1) hash table object registry.
- Add conservative register and stack root scanning with ASan-safe boundary traversal.
- Add deep DFS object, array, and closure traversal during mark phase.
- Add sweep phase reclaiming disconnected reference cycles and payload buffers.
- Add corpus test language/gc_cycle_collection.ts verifying cycle reclamation.

* feat(runtime): implement native RFC 6455 WebSocket client and strengthen tracing GC roots

- Add native RFC 6455 WebSocket client engine in internal/runtime/native/websocket/
  with HTTP 101 upgrade handshake, client-side 4-byte random masking, unmasking,
  ping/pong control frame handling, SO_NOSIGPIPE/MSG_NOSIGNAL, and non-blocking polling.
- Wire LLVM intrinsic lowering and runtime declarations for WebSocket APIs.
- Add TypeScript standard library contracts and adapter in stdlib/globals.d.ts and stdlib/ws.ts.
- Add corpus test internal/compiler/testdata/corpus/api/websocket_client.ts.
- Strengthen GC roots: register module global variable slots and scan execution stack
  to native stack bottom using pthread_get_stackaddr_np.
- Fix TypedArray slice GC registration cleanup and ArrayBuffer sweep layout.
- Synchronize docs/typescript-parity-report.md with WebSocket and Tracing GC statuses.

* fix(runtime): support WASI compilation in GC and silence 32-bit tautological warning

* fix(runtime): support WASI cross-compilation in WebSocket and guard runtime macros

* docs: synchronize WebSocket global in parity report

* test(corpus): use @native.expected for websocket and unify gc() output for Node parity

+++ internal/runtime/native/errors/runtime.c
@@ -231,7 +231,7 @@ int scriptgo_string_from_object(void *obj, char **out_str) {
         scriptgo_object_t *o = (scriptgo_object_t *)obj;
         if (o->magic == SCRIPTGO_OBJECT_MAGIC) {
             if (o->type_name != NULL && (strcmp(o->type_name, "Error") == 0 || strstr(o->type_name, "Error") != NULL)) {
-                if (o->field_count > 0 && o->fields[0] != 0 && o->fields[0] != 0x7FF8000000000000ULL) {
+                if (o->field_count > 0 && o->fields[0] != 0 && (uint64_t)o->fields[0] != 0x7FF8000000000000ULL) {
                     *out_str = strdup((const char *)o->fields[0]);
                     return 0;
                 }
